Cherry-pick PR's into the RB-2.5 branch for v2.5.5 release (#917)
* Revert "Disable OPENEXR_IMF_HAVE_GCC_INLINE_ASM_AVX when building on arm64 macOS"

This reverts commit 67053ebf1882ed93e74dbb7b5dcdcf0b2420c2c9.


* Fix Apple Universal 2 (arm64/x86_64) builds

* In these types of builds we want arm64 and x86_64 (with AVX optimisations).
  However the way cmake works (with `CMAKE_OSX_ARCHITECTURES="arm64;x86_64"`
  means that we share one OpenEXRConfigInternal.h between both builds. So
  we have to have OPENEXR_IMF_HAVE_GCC_INLINE_ASM_AVX mean "AVX GCC asm is
  available if platform is x86", rather than "AVX GCC asm is available".
  Then we decide on AVX optimisations based on that #define and also the
  platform defines.


* Include <limits> where required by newer compilers (#893)

* Include <limits> where required by newer compilers


* Removed redundant #include <limits>


* add buffer size validation to FastHuf decode


* prevent overflow in RgbaFile cachePadding


* Use size_t for DWA buffersize calculation (#901)

* Use size_t for DWA buffersize calculation


* use Int64 instead of size_t for buffersize calculations


* prevent overflows by using Int64 for all vars in DWA initialize (#903)


* update tileoffset sanitycheck to handle ripmaps  (#910)

* update tileoffset sanitycheck to handle ripmaps


* slight reorganization


* slight reorganization


* remove extra if statement from validateStreamSize


* additional verification of DWA data sizes (#914)


* Release notes for v2.5.5


* fix merge of ImfTiledInputFile.cpp


* Bump version for 2.5.5


* Only wait for and join joinable threads (#921)


* Fixed botched merge or IlmThread.cpp/IlmThreadPool.cpp


* Fix 2.5.5 release date

+++ OpenEXR/IlmImf/ImfDwaCompressor.h
@@ -169,16 +169,16 @@ class DwaCompressor: public Compressor
     std::vector<CscChannelSet> _cscSets;
     std::vector<Classifier>    _channelRules;
 
-    char             *_packedAcBuffer;
-    size_t            _packedAcBufferSize;
-    char             *_packedDcBuffer;
-    size_t            _packedDcBufferSize;
-    char             *_rleBuffer;
-    size_t            _rleBufferSize;
-    char             *_outBuffer;
-    size_t            _outBufferSize;
-    char             *_planarUncBuffer[NUM_COMPRESSOR_SCHEMES];
-    size_t            _planarUncBufferSize[NUM_COMPRESSOR_SCHEMES];
+    char*             _packedAcBuffer;
+    Int64             _packedAcBufferSize;
+    char*             _packedDcBuffer;
+    Int64             _packedDcBufferSize;
+    char*             _rleBuffer;
+    Int64             _rleBufferSize;
+    char*             _outBuffer;
+    Int64             _outBufferSize;
+    char*             _planarUncBuffer[NUM_COMPRESSOR_SCHEMES];
+    Int64             _planarUncBufferSize[NUM_COMPRESSOR_SCHEMES];
 
     Zip              *_zip;
     float             _dwaCompressionLevel;
